pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "switchbot_climate"
version = "0.1.0"
description = "A Python package to integrate SwitchBot devices with MQTT for climate control."
authors = [
    { name = "Aaron Watson", email = "watsona4@gmail.com" },
]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "paho-mqtt >= 2.1",
    "colorlog",
    "requests",
    "strictyaml",
]

[project.urls]
Homepage = "https://github.com/watsona4/switchbot_climate"

[project.optional-dependencies]
dev = [
    "black",
    "isort",
    "mypy",
    "types-requests",
]
test = [
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.sdist]
include = ["switchbot_climate/**", "README.md"]

[tool.hatch.build.targets.wheel]
include = ["switchbot_climate/**", "README.md"]

[tool.mypy]
check_untyped_defs = true
ignore_missing_imports = true
implicit_optional = true

[tool.black]
line-length = 99

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
pythonpath = ["."]
